update libs, improve layout, add dxmt patcher support

+++ Procyon/CXPatcher/Util.swift
@@ -6,14 +6,63 @@
 //
 import Foundation
 
+let PATCHED_CX_APPNAME = "Crossover_patched.app"
+let PATCHED_CX_X87_APPNAME = "Crossover_patched_x87.app"
 private let DEFAULT_CXP_BOTTLES_ROOTPATH = "/Users/${USER}/"
 private let DEFAULT_CXP_BOTTLES_FOLDER = "CXPBottles"
 private let DEFAULT_CXP_BOTTLES_PATH = DEFAULT_CXP_BOTTLES_ROOTPATH + DEFAULT_CXP_BOTTLES_FOLDER
 private let CROSSOVER_MAIN_CONFIGURATION = "/etc/CrossOver.conf"
 private let WINE_RESOURCES_ROOT = "Crossover"
 private let SHARED_SUPPORT_COMPONENT = "Contents/SharedSupport/CrossOver"
-private let SHARED_SUPPORT_PATH = "/" + SHARED_SUPPORT_COMPONENT
+let SHARED_SUPPORT_PATH = "/" + SHARED_SUPPORT_COMPONENT
 private let INFO_PLIST_PATH = "Contents/Info.plist"
+struct PathMap {
+    var src: String
+    var dst: String
+}
+private struct CXPlist: Decodable {
+    private enum CodingKeys: String, CodingKey {
+        case CFBundleIdentifier, CFBundleShortVersionString
+    }
+
+    let CFBundleIdentifier: String
+    let CFBundleShortVersionString: String
+}
+
+
+private let WINE_DXVK_RESOURCES_PATHS: [String] = [
+    "/lib/dxvk/i386-windows/d3d10core.dll",
+    "/lib/dxvk/i386-windows/d3d11.dll",
+    "/lib/dxvk/x86_64-windows/d3d10core.dll",
+    "/lib/dxvk/x86_64-windows/d3d11.dll",
+]
+
+let WINE_WINEINF_PATH: String = "/share/wine/wine.inf"
+
+
+let WINE_RESOURCES_PATHS: [String] = [
+    //    "/lib64/libinotify.0.dylib",
+    //    "/lib64/libinotify.dylib",
+    //    "/lib/wine/i386-windows/kernelbase.dll",
+    //    "/lib/wine/i386-windows/ntdll.dll",
+    //    "/lib/wine/i386-windows/winegstreamer.dll",
+    //    "lib/wine/i386-windows/wineboot.exe",
+    //    "/lib/wine/i386-windows/winecfg.exe",
+    //    "/lib/wine/x86_64-unix/ntdll.so",
+    //    "/lib/wine/x86_64-windows/kernelbase.dll",
+    //    "/lib/wine/x86_64-windows/ntdll.dll",
+    //    "/lib/wine/x86_64-windows/wineboot.exe",
+    //    "/lib/wine/x86_64-windows/winecfg.exe",
+//        "/share/crossover/bottle_data/crossover.inf",
+    //    "/CrossOver-Hosted Application/wineserver",
+]
+//+ [MOLTENVK_BASELINE]
+//+ [WINE_WINEINF_PATH]
+//+ WINE_DXVK_RESOURCES_PATHS
+//+ WINE_GSTREAMER_RESOURCES_PATHS
+//+ WINE_DXMT_RESOURCES_PATHS
+
+
 
 enum PatchMVK {
     case legacyUE4
@@ -51,7 +100,15 @@ private struct Env {
 
 private func disable(dest: String) {
     let f = FileManager.default
-    do {try f.moveItem(atPath: dest, toPath: dest  + "_disabled")
+    if f.fileExists(atPath: dest  + "_disabled") {
+        do {
+            try f.removeItem(atPath: dest  + "_disabled")
+        } catch {
+            console.error("can't remove file \(dest + "_disabled")")
+        }
+    }
+    do {
+        try f.moveItem(atPath: dest, toPath: dest  + "_disabled")
         console.log("disabling \(dest)")
     } catch {
         console.error("can't move file \(dest)")
@@ -62,11 +119,34 @@ func copyResource(name: String, destUrl: URL) throws {
     let f = FileManager.default
     if let resUrl = Bundle.main.url(forResource: name, withExtension: nil) {
         if(f.fileExists(atPath: destUrl.path())) {
-            try f.removeItem(at: destUrl)
+            let orig = destUrl.appendingPathExtension("orig")
+            if(!f.fileExists(atPath: orig.path())) {
+                try f.moveItem(at: destUrl, to: orig)
+            } else {
+                try f.removeItem(at: destUrl)
+            }
+            try f.copyItem(at: resUrl, to: destUrl)
+        } else {
+            console.error("Couldn't find \(name)")
         }
-        try f.copyItem(at: resUrl, to: destUrl)
+    }
+}
+
+func safeFileCopy(source: URL, dest: URL) throws {
+    let f = FileManager.default
+    if(f.fileExists(atPath: dest.path())) {
+        do {
+            try f.moveItem(at: dest, to: dest.appendingPathExtension("orig"))
+        } catch {
+            console.log(String(reflecting: error))
+        }
     } else {
-        console.error("Couldn't find \(name)")
+        console.log("file doesn't exist I'll just copy then")
+    }
+
+    do {
+        try f.copyItem(at: source, to: dest)
+        console.log("\(source) copied")
     }
 }
 
@@ -110,7 +190,7 @@ private func appendLinesToFile(fileURL: URL, additionalLines: [String]) -> Strin
 }
 
 private func getENVOverrideConfigfile(envs: [Env], fileURL: URL) -> String {
-    let additionallines: [String] = envs.map { env in
+    let additionallines: [String] = ["[EnvironmentVariables]"] + envs.map { env in
         toCrossoverENVString(env.key, env.value)
     }
     
@@ -130,25 +210,26 @@ private func addEnvs(_ envs: [Env], to: URL, from: URL) {
 
 func addGlobals(appURL: URL, opts: Opts) {
     disable(dest: appURL.path + SHARED_SUPPORT_PATH + CROSSOVER_MAIN_CONFIGURATION)
-    var envs: [Env] = [Env(key: "CX_BOTTLE_PATH", value: opts.cxbottlesPath)]
-    if(opts.patchMVK == .legacyUE4) {
-        console.log("add enable UE4 Hack env")
-        envs += [Env(key: "MVK_CONFIG_UE4_HACK_ENABLED", value: "1")]
-    }
-    if(opts.globalEnvs.dxvkAsync == true) {
-        console.log("add DXVK async env")
-        envs += [Env(key: "DXVK_ASYNC", value: "1")]
-    }
-    if(opts.globalEnvs.disableUE4Hack == true) {
-        console.log("add UE4 disable env")
-        envs += [Env(key: "NAS_DISABLE_UE4_HACK", value: "1")]
-    }
+    var envs: [Env] = [Env(key: "CX_BOTTLE_PATH", value: opts.cxbottlesPath)] // other envs to be added later
+//    envs += [Env(key: "NAS_DISABLE_UE4_HACK", value: "1")]
+//    if(opts.patchMVK == .legacyUE4) {
+//        console.log("add enable UE4 Hack env")
+//        envs += [Env(key: "MVK_CONFIG_UE4_HACK_ENABLED", value: "1")]
+//    }
+//    if(opts.globalEnvs.dxvkAsync == true) {
+//        console.log("add DXVK async env")
+//        envs += [Env(key: "DXVK_ASYNC", value: "1")]
+//    }
+//    if(opts.globalEnvs.disableUE4Hack == true) {
+//        console.log("add UE4 disable env")
+//        envs += [Env(key: "NAS_DISABLE_UE4_HACK", value: "1")]
+//    }
     if(opts.globalEnvs.disableMVKArgumentBuffers == true) { // to add the option later
         console.log("disable MoltenVK Argument Buffers")
         envs += [Env(key: "MVK_CONFIG_USE_METAL_ARGUMENT_BUFFERS", value: "0")]
     }
-    console.log("enable MoltenVK UE4 HAck")
-    envs += [Env(key: "MVK_CONFIG_UE4_HACK_ENABLED", value: "1")]
+//    console.log("enable MoltenVK UE4 Hack")
+//    envs += [Env(key: "MVK_CONFIG_UE4_HACK_ENABLED", value: "1")]
     
     addEnvs(envs, to: appURL.appendingPathComponent(SHARED_SUPPORT_PATH + CROSSOVER_MAIN_CONFIGURATION), from: appURL.appendingPathComponent(SHARED_SUPPORT_PATH + CROSSOVER_MAIN_CONFIGURATION + "_disabled"))
 }
@@ -162,46 +243,54 @@ func signAndFixup(destPath: String) throws {
     try fixup(destPath: destPath)
 }
 
-func makeX87CrossoverPatchedCopy (sourceCXPath: URL, patchedApp: URL) -> Void {
-    let f = FileManager.default
-    do {
-        let destUrl = f.homeDirectoryForCurrentUser.appendingPathComponent("Applications", isDirectory: true).appendingPathComponent("Crossover_x87.app")
-        // Make sure destination app doesn't exist and if it does, delete it
-        if (f.fileExists(atPath: destUrl.path())) {
-            try f.removeItem(at: destUrl)
-        }
-        // MARK: Step 1 copy the app in the user's application folder
-        try f.copyItem(at: sourceCXPath, to: destUrl)
-        
-        // MARK: Step 2 remove signature
-        if(f.fileExists(atPath: destUrl.path())) {
-            // Step 2 unsign app
-            try safeShell("codesign --remove-signature \"\(destUrl.path())\"")
-            try safeShell("codesign --remove-signature \"\(destUrl.appendingPathComponent("Contents/SharedSupport/CrossOver/lib/wine/x86_64-unix/wine").path())\"")
-            
-            // MARK: Step 3 fix app after patching
-            try fixup(destPath: destUrl.path())
-        } else {
-            console.error("Couldn't find Crossover_x87.app in \(destUrl.path())")
-        }
-    } catch {
-        console.error(String(reflecting: error))
+func removeSignature(destURL: URL) throws {
+    try safeShell("codesign --remove-signature \"\(destURL.path())\"")
+    let command = "codesign --remove-signature \"\(destURL.appendingPathComponent("Contents/SharedSupport/CrossOver/lib/wine/x86_64-unix/wine").path())\""
+    try safeShell(command)
+}
+
+private func parseCXPlist(plistPath: String) -> CXPlist {
+    let data = try! Data(contentsOf: URL(filePath: plistPath))
+    let decoder = PropertyListDecoder()
+    return try! decoder.decode(CXPlist.self, from: data)
+}
+
+private func markAsPatched(url: URL) {
+    let plist = parseCXPlist(plistPath: url.path + "/Contents/Info.plist")
+    if let version = Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String {
+        editInfoPlist(at: url, key: "CFBundleShortVersionString", value: plist.CFBundleShortVersionString + "p" + version)
     }
 }
 
+func makeX87CrossoverPatchedCopy (sourceCXPath: URL, patchedApp: URL) async -> Void {
+    await makeCrossoverPatchedCopy(sourceCXPath: sourceCXPath, setProgress: { _ in }, setLoading: { _ in }, isX87: true)
+}
+
+func fetchLatestRelease(from path: String) async throws -> String {
+    let url = URL(string: "\(path)/releases/latest")!
+    let (data, _) = try await URLSession.shared.data(from: url)
+    let json = try JSONSerialization.jsonObject(with: data) as! [String: Any]
+    return json["tag_name"] as! String
+}
 
 
-func makeCrossoverPatchedCopy (sourceCXPath: URL, setProgress: @escaping (Double) -> Void, setLoading: @escaping (Bool) -> Void) async -> URL {
+@discardableResult
+func makeCrossoverPatchedCopy(sourceCXPath: URL, setProgress: @escaping (Double) -> Void, setLoading: @escaping (Bool) -> Void, isX87: Bool = false) async -> URL {
     let f = FileManager.default
-    let name = "Crossover_patched.app"
-    let ROOT = "Contents/SharedSupport/"
+    let name = isX87 ? PATCHED_CX_X87_APPNAME : PATCHED_CX_APPNAME
     let destUrl = f.homeDirectoryForCurrentUser.appendingPathComponent("Applications", isDirectory: true).appendingPathComponent(name)
     let resources: [(res: String, dest: URL)] = [
-        (res: "ntdll.so", dest: "CrossOver/lib/wine/x86_64-unix/"),
-        (res: "winedmo.so", dest: "CrossOver/lib/wine/x86_64-unix/"),
-        (res: "winegstreamer.so", dest: "CrossOver/lib/wine/x86_64-unix/"),
-    ].map { item in
-        (res: item.res, dest: destUrl.appendingPathComponent(ROOT + item.dest + item.res))
+        (res: "ntdll.so", dest: "/lib/wine/x86_64-unix/ntdll.so"),
+        (res: "winedmo.so", dest: "/lib/wine/x86_64-unix/winedmo.so"),
+        (res: "d9vk/x32/d3d9.dll", dest: "/lib/wine/i386-windows/d3d9.dll"),
+        (res: "d9vk/x64/d3d9.dll", dest: "/lib/wine/x86_64-windows/d3d9.dll"),
+        (res: "winegstreamer.so", dest: "/lib/wine/x86_64-unix/winegstreamer.so"),
+//        (res: "libMoltenVK-experimental.dylib", dest: "/lib64/libMoltenVK.dylib"),
+    ].compactMap { item in
+        (res: item.res, dest: destUrl.appendingPathComponent(SHARED_SUPPORT_COMPONENT + item.dest))
+    }.filter {
+        resource in
+        isX87 ? true : !resource.res.contains("d3d9")
     }
     do {
         // Make sure destination app doesn't exist and if it does, delete it
@@ -213,38 +302,23 @@ func makeCrossoverPatchedCopy (sourceCXPath: URL, setProgress: @escaping (Double
         
         if(f.fileExists(atPath: destUrl.path())) {
             // MARK: Step 1 copy resources
-            for (res, dest) in resources{
+            for (res, dest) in resources {
                 console.log("Copying \(res) to \(dest.path())")
                 try copyResource(name: res, destUrl: dest)
             }
             // MARK: Step 2 download gstreamer
             let gstURL = try await getGstreamerDownloadURL()
+            let (url: dxmtURL, versionTag: dxmtVersion) = try await getDXMTDownloadURL()
             console.log("Gstreamer download url: \(gstURL)")
-            try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<Void, Error>) in
+            try await withCheckedThrowingContinuation { continuation in
                 let gstreamerDownloader = TarDownloader(
                     fromUrl: gstURL,
                     onProgress: { progress in
                         setProgress(progress)
                     },
-                    onComplete: { url in
+                    onComplete: { srcUrl in
                         do {
-                            let src = url.appendingPathComponent("GStreamer.framework")
-                            let dst = destUrl.appendingPathComponent("Contents/SharedSupport/CrossOver/lib64/")
-                            if f.fileExists(atPath: dst.appendingPathComponent("GStreamer.framework/.gitignore").path) { // need to remove this
-                                try f.removeItem(at: dst.appendingPathComponent("GStreamer.framework/.gitignore"))
-                            }
-                            try f.copyItem(at: src, to: dst.appendingPathComponent("GStreamer.framework"))
-                            try f.removeItem(at: dst.appendingPathComponent("gstreamer-1.0"))
-                            let opts = Opts()
-                            // MARK: Step 3 add env variables to crossover configuration
-                            addGlobals(appURL: destUrl, opts: opts)
-                            // MARK: Step 4 disable auto update
-                            if(opts.autoUpdateDisable) {
-                                disableAutoUpdate(url: destUrl)
-                            }
-                            // MARK: Step 5 sign
-                            // MARK: Step 6 fix app after patching
-                            try signAndFixup(destPath: destUrl.path())
+                            try installGstreamer(srcUrl: srcUrl, destUrl: destUrl )
                             continuation.resume()
                         } catch {
                             console.error(String(reflecting: error))
@@ -253,6 +327,7 @@ func makeCrossoverPatchedCopy (sourceCXPath: URL, setProgress: @escaping (Double
                         setLoading(false)
                     },
                     onError: { error in
+                        console.error("Error while downloading GStreamer")
                         console.error(String(reflecting: error))
                         continuation.resume(throwing: error)
                     }
@@ -260,8 +335,48 @@ func makeCrossoverPatchedCopy (sourceCXPath: URL, setProgress: @escaping (Double
                 setLoading(true)
                 gstreamerDownloader.download()
             }
+            try await withCheckedThrowingContinuation { continuation in
+                let dxmtDownloader = TarDownloader(
+                    fromUrl: dxmtURL,
+                    onProgress: { progress in
+                        setProgress(progress)
+                    },
+                    onComplete: { srcURL in
+                        do {
+                            try installDXMT(srcURL: srcURL, destUrl: destUrl, versionTag: dxmtVersion)
+                            continuation.resume()
+                        } catch {
+                            console.error(String(reflecting: error))
+                            continuation.resume(throwing: error)
+                        }
+                        setLoading(false)
+                    },
+                    onError: { error in
+                        console.error("Error while downloading DXMT")
+                        console.error(String(reflecting: error))
+                        continuation.resume(throwing: error)
+                    }
+                )
+                setLoading(true)
+                dxmtDownloader.download()
+            }
+            let opts = Opts()
+            // MARK: Step 3 add env variables to crossover configuration
+            addGlobals(appURL: destUrl, opts: opts)
+            // MARK: Step 4 disable auto update
+            if(opts.autoUpdateDisable) {
+                disableAutoUpdate(url: destUrl)
+            }
+            markAsPatched(url: destUrl)
+            // MARK: Step 5 sign
+            // MARK: Step 6 fix app after patching
+            if !isX87 {
+                try signAndFixup(destPath: destUrl.path())
+            } else {
+                try removeSignature(destURL: destUrl)
+            }
         } else {
-            console.error("Couldn't find Crossover_x87.app in \(destUrl.path())")
+            console.error("Couldn't find Crossover app in \(destUrl.path())")
         }
     } catch {
         console.error(String(reflecting: error))
